requirements/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async==0.6.2
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-log==0.4.0
colorama==0.4.6
coloredlogs==15.0.1
Crawl4AI==0.7.4
cryptography==44.0.2
cssselect==1.3.0
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docling==2.28.0
docling-core==2.23.3
docling-ibm-models==3.4.1
docling-parse==4.0.0
easyocr==1.7.2
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fake-http-header==0.3.5
fake-useragent==2.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastembed==0.5.1
faust-cchardet==2.1.19
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
humanize==4.12.1
hyperframe==6.1.0
identify==2.6.9
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-docling==0.2.0
langchain-experimental==0.3.4
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langgraph==0.2.71
langgraph-checkpoint==2.0.12
langgraph-checkpoint-mongodb==0.1.0
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.51
langsmith==0.3.8
lark==1.2.2
latex2mathml==3.77.0
lazy_loader==0.4
litellm==1.63.6
loguru==0.7.3
lxml==5.3.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==4.1.0
motor==3.7.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.2.4
onnxruntime==1.20.1
openai==1.68.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
patchright==1.52.5
pillow==10.4.0
playwright==1.54.0
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
py_rust_stemmers==0.1.5
pyasn1==0.4.8
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pymongo==4.9.2
PyMuPDF==1.25.4
pymupdf4llm==0.0.17
pyOpenSSL==25.0.0
pytesseract==0.3.13
pypdfium2==4.30.1
pyperclip==1.9.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
qdrant-client==1.13.3
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.1
router==0.1
routers==0.10.1
rpds-py==0.23.1
rsa==4.9
rtree==1.4.0
ruff==0.11.2
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
semchunk==2.2.2
sentence-transformers==4.0.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tf-playwright-stealth==1.1.2
threadpoolctl==3.5.0
tifffile==2025.3.13
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
trimesh==4.7.4
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
#uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0